+++ src/module/disqus.model.ts
@@ -1,0 +1,7 @@
+export interface DisqusComment {
+  id: number;
+  name: string;
+}
+export interface DisqusReady {
+  height: number;
+}