+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 2.2.0
+
+ - Remove `@Optional` from DisqusService, closes #18.
+ - feat(onReady, onPaginate) events.
+ - rename `comment` output to `onNewComment`.
+
 ## 2.1.1
 
  - support for Universal